Add create/update endpoints for applications, and setWeekData/getWeekData for application data
- Renamed applications list endpoint
- Cleanup migrations and phpunit config for testing db with sqlite
- Add Api\Application functional tests
- Fix migrations so they work with both mysql and sqlite
- Add seeders for common db tables

+++ src/database/migrations/2015_11_21_311070_create_tracker_paths_table.php
@@ -20,6 +20,11 @@ class CreateTrackerPathsTable extends Migration {
 	 */
 	public function up()
 	{
+        // Do not run this migration in testing env
+        if (!env('TRACKER_ENABLED')) {
+            return;
+        }
+
         Schema::connection('tracker')->create($this->table, function (Blueprint $table) {
             $table->bigIncrements('id');
 
@@ -37,7 +42,12 @@ class CreateTrackerPathsTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop($this->table);
+        // Do not run this migration in testing env
+        if (env('APP_ENV') == 'testing') {
+            return;
+        }
+
+        Schema::drop($this->table);
 	}
 
 }
