+++ src/database/migrations/2015_11_21_311095_add_tracker_referer_columns.php
@@ -3,25 +3,31 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class AddTrackerRefererColumns extends Migration {
+class AddTrackerRefererColumns extends Migration
+{
 
-	/**
-	 * Table related to this migration.
-	 *
-	 * @var string
-	 */
+    /**
+     * Table related to this migration.
+     *
+     * @var string
+     */
 
-	private $table = 'tracker_referers';
+    private $table = 'tracker_referers';
 
-	private $foreign = 'tracker_referers_search_terms';
+    private $foreign = 'tracker_referers_search_terms';
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        // Do not run this migration in testing env
+        if (!env('TRACKER_ENABLED') || Schema::hasColumn($this->table, 'medium')) {
+            return;
+        }
+
         Schema::connection('tracker')->table($this->table, function (Blueprint $table) {
             $table->string('medium')->nullable()->index();
             $table->string('source')->nullable()->index();
@@ -29,33 +35,37 @@ class AddTrackerRefererColumns extends Migration {
         });
 
         Schema::connection('tracker')->table($this->foreign, function (Blueprint $table) {
-			$table->foreign('referer_id', 'tracker_referers_referer_id_fk')
-				->references('id')
-				->on('tracker_referers')
-				->onUpdate('cascade')
-				->onDelete('cascade');
-		});
-	}
+            $table->foreign('referer_id', 'tracker_referers_referer_id_fk')
+                  ->references('id')
+                  ->on('tracker_referers')
+                  ->onUpdate('cascade')
+                  ->onDelete('cascade');
+        });
+    }
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-        Schema::connection('tracker')->table($this->table, function (Blueprint $table) {
-				$table->dropColumn('medium');
-				$table->dropColumn('source');
-				$table->dropColumn('search_terms_hash');
-			}
-		);
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        // Do not run this migration in testing env
+        if (env('APP_ENV') == 'testing') {
+            return;
+        }
 
         Schema::connection('tracker')->table($this->table, function (Blueprint $table) {
-				$table->dropForeign('tracker_referers_referer_id_fk');
-			}
-		);
+            $table->dropColumn('medium');
+            $table->dropColumn('source');
+            $table->dropColumn('search_terms_hash');
+        }
+        );
 
-	}
+        Schema::connection('tracker')->table($this->table, function (Blueprint $table) {
+            $table->dropForeign('tracker_referers_referer_id_fk');
+        }
+        );
+    }
 
 }
